Refactor test utilities: rename package from testing to testhelpers and update imports
