+++ internal/infrastructure/category/category_repository_test.go
@@ -4,13 +4,14 @@ import (
 	"context"
 	"testing"
 
+	testutils "family-budget-service/internal/testhelpers"
+
 	"github.com/google/uuid"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 
 	"family-budget-service/internal/domain/category"
 	categoryrepo "family-budget-service/internal/infrastructure/category"
-	testutils "family-budget-service/internal/testing"
 )
 
 func TestCategoryRepositoryPostgreSQL_Integration(t *testing.T) {
