+++ internal/infrastructure/transaction/transaction_repository_test.go
@@ -5,13 +5,14 @@ import (
 	"testing"
 	"time"
 
+	testutils "family-budget-service/internal/testhelpers"
+
 	"github.com/google/uuid"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 
 	"family-budget-service/internal/domain/transaction"
 	transactionrepo "family-budget-service/internal/infrastructure/transaction"
-	testutils "family-budget-service/internal/testing"
 )
 
 func TestTransactionRepositoryPostgreSQL_Integration(t *testing.T) {
